Add placeholder logic for batch ID

+++ legend-engine-xts-persistence/legend-engine-xt-persistence-component/legend-engine-xt-persistence-component-relational-core/src/main/java/org/finos/legend/engine/persistence/components/relational/api/DatasetIngestResultsAbstract.java
@@ -31,7 +31,7 @@ public abstract class DatasetIngestResultsAbstract
 {
     public abstract Optional<String> ingestRequestId();
 
-    public abstract int batchId();
+    public abstract long batchId();
 
     public abstract String dataset();
